requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pyrebase4==4.7.1
firebase-admin==6.5.0
pyjwt==2.8.0
